app/app by kimi/ml model/requirements.txt
groq
python-dotenv
PyPDF2
python-docx
requests
fastapi
uvicorn[standard]
python-multipart
Pillow
pytesseract
PyMuPDF
